pyproject.toml
[project]
name = "doctr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.45",
    "typer>=0.16.1",
]
